fix: fixed bugs preventing proper env reset
After each episode (generation of full summary for given article/summary pair), the environment is meant to reset and the dataloader yield a new article/summary pair. The data loader iterator was broken due to a missing incrementer. The env step function was broken due to a missing reset() call. The reset() method was broken due to a missing data iterator reinstantiation on StopIteration. This commit fixes these three issues.